Move dashboard page under overview
